requirements.txt
chromadb>=0.5.5
sentence-transformers>=3.2.0
google-genai>=0.3.0
pypdf>=4.0.1
pydantic>=2.9.0
tenacity>=8.2.3
numpy>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.0